% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LinearExample}
\alias{LinearExample}
\title{Simulated data for functional linear regression.}
\format{
List containing the following elements:
\describe{
\item{X.curves}{200 by 101 matrix of functional predictors.}
\item{Y}{200 by 1 numeric vector of Gaussian responses.}
\item{beta.true}{The true functional parameter \eqn{\beta}.}
}
}
\usage{
data(LinearExample)
}
\description{
Randomly generated data with Gaussian responses for functional linear regression example follows Case I from Liu et. al. (2022).
}
\references{
Xi Liu, Afshin A. Divani, and Alexander Petersen. "Truncated estimation in functional generalized linear regression models" (2022). \emph{Computational Statistics & Data Analysis}.
}
\keyword{data}
