% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FGLMtrunc.R
\name{plot.FGLMtrunc}
\alias{plot.FGLMtrunc}
\title{Plot functional parameters \eqn{\beta} from a \code{FGLMtrunc} object}
\usage{
\method{plot}{FGLMtrunc}(x, include_smooth = TRUE, ...)
}
\arguments{
\item{x}{fitted \code{FGLMtrunc} object}

\item{include_smooth}{If TRUE, smoothing estimate without truncation of \eqn{\beta} is plotted.}

\item{\dots}{additional plot arguments}
}
\value{
No return value.
}
\description{
Plot functional parameters \eqn{\beta} as a function of \eqn{t} for a fitted
\code{FGLMtrunc} object.
}
