% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.truncatedFGLM.R
\name{plot.truncatedFGLM}
\alias{plot.truncatedFGLM}
\title{Plot functional parameters \eqn{\beta} from a \code{truncatedFGLM} object}
\usage{
\method{plot}{truncatedFGLM}(x, include_smooth = TRUE, ...)
}
\arguments{
\item{x}{fitted \code{truncatedFGLM} object}

\item{include_smooth}{If TRUE, smoothing estimate without truncation of \eqn{\beta} is plotted.}

\item{\dots}{additional plot arguments}
}
\description{
Plot functional parameters \eqn{\beta} as a function of \eqn{t} for a fitted
\code{truncatedFGLM} object.
}
