% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FGLMtrunc.R
\name{print.FGLMtrunc}
\alias{print.FGLMtrunc}
\title{Print a \code{FGLMtrunc} object}
\usage{
\method{print}{FGLMtrunc}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted \code{FGLMtrunc} object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\value{
The fitted object is silently return.
}
\description{
Print a summary of truncation point of the fitted \code{FGLMtrunc} model.
}
\details{
Truncation point estimate of \eqn{\delta} is printed.
}
