% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.truncatedFGLM.R
\name{print.truncatedFGLM}
\alias{print.truncatedFGLM}
\title{Print a \code{truncatedFGLM} object}
\usage{
\method{print}{truncatedFGLM}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted \code{truncatedFGLM} object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\description{
Print a summary of truncation point of the fitted \code{truncatedFGLM} model.
}
\details{
Truncation point estimate of \eqn{\delta} is printed.
}
